Change: update project structure - Upgrade to Gradle 7.5 - Add ASDF `.tool-versions` and standardize temurin JDK for Java 11 - Remove the `github` part from packaging
