+++ build.gradle
@@ -1,15 +1,16 @@
-apply plugin: 'java'
-apply plugin: 'maven'
-
-version='0.0.6'
+plugins {
+    id 'java'
+}
 
 repositories {
-  mavenCentral()
+    mavenCentral()
 }
 
 dependencies {
-  testCompile 'junit:junit:4.12'
-  testCompile 'org.assertj:assertj-core:3.4.0'
-  testCompile 'org.openjdk.jmh:jmh-core:1.9.3'
-  testCompile 'org.openjdk.jmh:jmh-generator-annprocess:1.9.3'
+    testImplementation 'org.junit.jupiter:junit-jupiter:5.8.2'
+    testImplementation 'org.hamcrest:hamcrest:2.2'
+}
+
+tasks.named('test') {
+    useJUnitPlatform()
 }
