+++ src/main/java/com/ggalmazor/ltdownsampling/Point.java
@@ -1,4 +1,4 @@
-package com.github.ggalmazor.ltdownsampling;
+package com.ggalmazor.ltdownsampling;
 
 import java.math.BigDecimal;
 import java.math.MathContext;
