+++ src/main/java/com/ggalmazor/ltdownsampling/tools/SlidingCollector.java
@@ -1,4 +1,4 @@
-package com.github.ggalmazor.ltdownsampling.tools;
+package com.ggalmazor.ltdownsampling.tools;
 
 import java.util.*;
 import java.util.function.BiConsumer;
