+++ src/test/resources/daily-foreign-exchange-rates-31-.csv
